% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_net.R
\name{u_net}
\alias{u_net}
\title{Create a U-Net architecture}
\usage{
u_net(
  net_h,
  net_w,
  grayscale = FALSE,
  layers_per_block = 2,
  blocks = 4,
  n_class = 1,
  filters = 16,
  dropout = 0,
  batch_normalization = TRUE,
  kernel_initializer = "he_normal"
)
}
\arguments{
\item{net_h}{Input layer height.}

\item{net_w}{Input layer width.}

\item{grayscale}{Defines input layer color channels - 1 if `TRUE`, 3 if `FALSE`.}

\item{layers_per_block}{Number of convolutional layers per block (can be 2 or 3)}

\item{blocks}{Number of blocks in the model.}

\item{n_class}{Number of classes.}

\item{filters}{Integer, the dimensionality of the output space (i.e. the number of output filters in the convolution).}

\item{dropout}{Dropout rate.}

\item{batch_normalization}{Should batch normalization be used in the block.}

\item{kernel_initializer}{Initializer for the kernel weights matrix.}
}
\value{
U-Net model.

A keras model
}
\description{
Create a U-Net architecture
}
\details{
This function was adapted and slightly modified from the u_net() function in the platypus package (\url{https://github.com/maju116/platypus/blob/master/R/u_net.R}).

Differences compared to platypus implementation:

- added argument: layers_per_block (can be 2 or 3)  \cr
- kernel size in layer_conv_2d_transpose is 2, not 3.  \cr
- dropout layers are only included if user specifies dropout > 0 \cr
- n_class = 1 by default (sufficient for binary classification used for vegetation)  \cr
- activation of output layer: "sigmoid" if n_class = 1, otherwise "softmax" \cr
- no checks: allows non-square input (e.g. 160x256)  \cr
}
